Paying for my sins...
I was developing this project in both a VM and my host machine at the
same time.
Unfortunately, the system time on the VM was completely wrong.
This caused the commits to have all kinds of out of order dates.

This is an uber squashed commit of the entire project.
I'm too lazy to go back and fix the dates, and history is of marginal
use here.